light refactoring ahead of lib tests
